agents/backend/onyx/server/features/email_sequence/requirements.txt
# Email Sequence Module - LangChain Integration
# Core dependencies

# LangChain and AI
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0

# Data validation and serialization
pydantic>=2.0.0
pydantic[email]>=2.0.0

# Async HTTP and email
aiohttp>=3.8.0
aiosmtplib>=2.0.0

# Database and ORM (optional)
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.28.0

# Web framework (optional)
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Monitoring and logging
structlog>=23.1.0
prometheus-client>=0.17.0

# Optional: Redis for caching
redis>=4.6.0
aioredis>=2.0.0

# Optional: Celery for background tasks
celery>=5.3.0

# Optional: Email providers
sendgrid>=6.10.0
mailgun>=0.1.0

# Core PyTorch and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-geometric>=2.3.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0

# Mixed Precision Training
torch-amp>=0.1.0
apex>=0.1.0

# Performance Optimization and Profiling
psutil>=5.9.0
memory-profiler>=0.61.0
line-profiler>=4.1.0
py-spy>=0.3.0
pyinstrument>=4.6.0
cProfile-to-ps>=0.1.0
snakeviz>=2.1.0

# Data Processing and Loading
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
h5py>=3.8.0
lmdb>=1.4.0
msgpack>=1.0.5
msgpack-numpy>=0.4.8

# Visualization and Plotting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0
altair>=5.1.0

# Web Interface and Gradio
gradio>=3.35.0
streamlit>=1.25.0
dash>=2.11.0
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Database and Storage
sqlalchemy>=2.0.0
alembic>=1.11.0
redis>=4.6.0
pymongo>=4.4.0
elasticsearch>=8.8.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0
toml>=0.10.2
configparser>=5.3.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Logging and Monitoring
loguru>=0.7.0
structlog>=23.1.0
prometheus-client>=0.17.0
grafana-api>=1.0.3

# Testing and Quality Assurance
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
coverage>=7.2.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Async and Concurrency
asyncio-mqtt>=0.13.0
aiofiles>=23.1.0
aiohttp>=3.8.5
websockets>=11.0.3

# Text Processing and NLP
nltk>=3.8.1
spacy>=3.6.0
textblob>=0.17.1
gensim>=4.3.0
wordcloud>=1.9.2
jieba>=0.42.1

# Email and Communication
smtplib2>=0.2.1
email-validator>=2.0.0
mailchimp-marketing>=3.0.75
sendgrid>=6.10.0

# API and Web Services
requests>=2.31.0
httpx>=0.24.0
aiohttp>=3.8.5
openai>=0.28.0
anthropic>=0.7.0
langchain>=0.0.267
langchain-community>=0.0.10

# File and Data Handling
pathlib2>=2.3.7
watchdog>=3.0.0
filelock>=3.12.0
tqdm>=4.65.0
rich>=13.4.0
click>=8.1.0

# Security and Authentication
cryptography>=41.0.0
bcrypt>=4.0.1
passlib>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.6

# System and OS Utilities
shutil>=1.0.0
subprocess32>=3.5.4
psutil>=5.9.0
GPUtil>=1.4.0
nvidia-ml-py3>=7.352.0

# Optimization and Hyperparameter Tuning
optuna>=3.2.0
hyperopt>=0.2.7
ray[tune]>=2.6.0
wandb>=0.15.0
mlflow>=2.5.0
tensorboard>=2.13.0

# Data Validation and Schema
pydantic>=2.0.0
marshmallow>=3.20.0
cerberus>=1.3.5
jsonschema>=4.19.0

# Caching and Performance
redis>=4.6.0
diskcache>=5.6.0
joblib>=1.3.0
cachetools>=5.3.0

# Error Handling and Debugging
sentry-sdk>=1.28.0
better-exceptions>=0.3.0
icecream>=2.1.3
debugpy>=1.6.7

# Development and Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0
jupyter>=1.0.0
jupyterlab>=4.0.0
ipython>=8.14.0

# Version Control and Deployment
gitpython>=3.1.31
docker>=6.1.0
kubernetes>=27.2.0

# Additional Utilities
python-dateutil>=2.8.2
pytz>=2023.3
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
more-itertools>=10.1.0
toolz>=0.12.0

# GPU and CUDA Support (Optional)
# Uncomment if using CUDA
# cupy-cuda11x>=12.0.0
# cupy-cuda12x>=12.0.0

# Distributed Training (Optional)
# Uncomment if using distributed training
# torch-distributed>=2.0.0
# horovod>=0.28.0

# Cloud and AWS Support (Optional)
# Uncomment if using AWS
# boto3>=1.28.0
# s3fs>=2023.6.0

# Google Cloud Support (Optional)
# Uncomment if using Google Cloud
# google-cloud-storage>=2.10.0
# google-cloud-aiplatform>=1.35.0

# Azure Support (Optional)
# Uncomment if using Azure
# azure-storage-blob>=12.17.0
# azure-ai-ml>=1.8.0 